requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.4
openpyxl==3.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
waitress==3.0.0
Werkzeug==3.0.4
